requirements.txt
Flask==2.3.3
gunicorn==20.1.0
python-pptx==0.6.21
PyPDF2==1.26.0
google-generativeai==0.8.3